requirements.txt
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.5
django-allauth==0.42.0
django-appconf==1.0.4
django-ckeditor==5.9.0
django-embed-video==1.3.3
django-imagekit==4.0.2
django-js-asset==1.2.2
django-markdownx==3.0.1
django-mdeditor==0.1.18
ExifRead==2.3.1
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
oauthlib==3.1.0
pilkit==2.0
Pillow==7.2.0
Pygments==2.6.1
pymdown-extensions==7.1
PyMySQL==0.10.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10
zipp==3.1.0
